Report application initialization statistics
I wanted it to be shown in bootstrap console, but it looks like by
this time, the stdout/stderr handles are redirected (not teed) to
Spyder Internal Console
